Minimal example showing how to submit colors directly from Viz Palette
When you click the Viz Palette link from the Data Explorer's color picker, the page will open in a new tab. Open up devtools from the Viz Palette page and paste in the script from the gist below. (If this turns out to be useful feature we want to support, we would have to implement some version of this on the Viz Palette site)

https://gist.github.com/BenRussert/0d35366810c11566489876e40155b1e8